chore: refactor SandoboxExample model
