+++ src/model/themes.ts
@@ -43,3 +43,5 @@ export const $themeName = createStore<ThemeNamesType>('default')
 $theme.on(changeThemeEvent, (_, themeName) => $themes.getState()[themeName])
 
 $themeName.on(changeThemeEvent, (_, themeName) => themeName)
+
+$theme.watch((theme) => console.log(theme))
