+++ src/model/tokens.ts
@@ -1,12 +1,9 @@
-import { createStore, createEvent, createEffect, forward, combine, Store } from 'effector'
+import { createStore, createEvent, createEffect, forward, Store } from 'effector'
 import { createGate } from 'effector-react'
 import { toast } from 'react-toastify'
-import deepmerge from 'deepmerge'
-import YAML from 'yaml'
 
 import { downloadTokens } from '../api/downloadTokens'
 import { getQueryParameter } from '../utils/queryParameters'
-import { toDeepToken } from '../utils/toDeepToken'
 import { changeThemeEvent } from './themes'
 import { updateTokensQueryParameterEvent } from './query'
 import { VariablesType, ThemeNamesType } from '../types'
@@ -20,33 +17,16 @@ export const uploadTokensEvent = createEvent()
 
 export const loadingTokensEvent = createEvent<boolean>()
 
-export const $cssVariables = createStore({})
 export const $designTokens = createStore<any>({})
 
-export const $cssText = combine(
-  { cssVariables: $cssVariables, designTokens: $designTokens },
-  ({ cssVariables, designTokens }) => {
-    const cssText = Object.keys(cssVariables).reduce((acc: string, v: string) => {
-      if (designTokens[v.replace('--', '')].changed) {
-        //@ts-ignore
-        acc += `  ${v}: ${cssVariables[v]};\n`
-      }
-      return acc
-    }, '')
-    return `:root {\n${cssText}}`
-  },
-)
-
-export const $yamlText = combine({ designTokens: $designTokens }, ({ designTokens }) => {
-  const yml = Object.entries(designTokens).reduce((acc, value: any) => {
-    if (value[1].changed) {
-      acc.push(toDeepToken(value[1].path, { value: value[1].value }))
-    }
-    return acc
-  }, [] as any)
-
-  const deepmergedYml = deepmerge.all(yml)
-  return YAML.stringify(deepmergedYml)
+export const $cssVariables = $designTokens.map((tokens) => {
+  return Object.values<VariablesType>(tokens).reduce(
+    (acc, { name, value }) => ({
+      ...acc,
+      [`--${name}`]: value,
+    }),
+    {},
+  )
 })
 
 export const $listDesignTokens: Store<VariablesType[]> = $designTokens.map((tokens) =>
@@ -90,17 +70,8 @@ $designTokens
   })
   .reset(variablesResetEvent)
 
-// TODO: Удалять значение из стора если change=false.
-$cssVariables
-  .on(variablesChangedEvent, (state, { name, value }) => ({ ...state, [`--${name}`]: value }))
-  .on(variablesChangedBatchEvent, (state, tokens) => {
-    const ret: Record<string, any> = {}
-    tokens.forEach((v) => (ret[`--${v.name}`] = v.value))
-    return { ...state, ...ret }
-  })
-  .reset(variablesResetEvent)
-
 variablesResetEvent.watch(() => updateTokensQueryParameterEvent())
+$designTokens.watch((tokens) => console.log(tokens))
 
 forward({
   from: variablesInitializationGate.open,
